fix(ci): set default working directory for all jobs
This commit fixes a `ERR_PNPM_NO_LOCKFILE` error in the CI workflows.

The error was caused by commands being executed in the root of the workspace directory instead of the subdirectory where the code is checked out.

This fix adds a `defaults.run.working-directory` to all jobs to ensure commands are run in the correct location.